+++ src/compile/tolk/config.ts
@@ -3,5 +3,6 @@ export type TolkCompilerConfig = {
     entrypoint: string;
     optimizationLevel?: number;
     withStackComments?: boolean;
+    withSrcLineComments?: boolean;
     experimentalOptions?: string;
 };